Non-functioning first commit
Only the stubs are here and the code is horrible, but better to get code in now.
The scrach.rb has a working CBST -> sorted array function and sorted array ->
CBST function that I can use to build the C versions of add and remove and then
start writing the benchmarking suite!